Added the Input Core.  Works well but realized SDL renderer is not able to
render at floating point precision. It leads to jittery movement. Result
is acceptable if ratio between rendertime and animation offset is "good".
Yet to find out what  "good" is. Animation is tied to framerate at the moment, which is bad, too.
A few more abstactions to come as well as the Audio Core and a XML loader for maps then it will be ready for
the Jam. The whole thing is a hack but that is part of the fun :)