Two players can run around the screen. Both share the same spritesheet but are individually controllable
and have individual properties.
Possible Implementation of GameObject added: Player.

Todos:
- think about where to do the updates for game objects. At the moment it is
a hybrid between the main and the actual game object implementation (in this case the player).

- The rendering is still weak. Find a way to decouple rendering and updates.

- Resource management is not available. There are absolute paths in the source. lol

- Maybe introduce a component system. At the moment there would be the need
for inheritance for specializations of GameObject

----------------------

It would be nice if all the mentioned todos above could be solved but the jam is a hack anyway
so probably we're good with deep inheritance for GameObjects. Big nono normally.