Map's resources get freed if load is called.
Some renaming of (member)variables.

Still a lot of memory that does not get freed properly. But this is probably
as far as it gets with the cleanup. goto is still there. Well, it is not
beautiful but contributes to the hacky approach so it kinda fits. Time to publish on gamejolt :)